Move source into lib directory
