+++ .github/workflows/node.js.yml
@@ -29,6 +29,7 @@ jobs:
           cache: pnpm
       - name: Install dependencies
         run: pnpm install
+      # Don't use npm run ci so that we only lint on an OS with sane CRLF.
       - name: Check coding standards
         if: matrix.os == 'ubuntu-latest'
         run: npm run lint
