+++ package.json
@@ -2,8 +2,9 @@
   "name": "@peggyjs/from-mem",
   "version": "2.0.3",
   "description": "Load a string as if it was an esm or commonjs module with a given filename.",
-  "main": "index.js",
+  "main": "./lib/index.js",
   "scripts": {
+    "clean": "rimraf coverage docs lib/*.d.ts",
     "build": "tsc",
     "lint": "eslint .",
     "test": "c8 node --stack-trace-limit=1000 --experimental-vm-modules --test",
@@ -30,6 +31,7 @@
     "@types/semver": "7.7.0",
     "c8": "10.1.3",
     "eslint": "9.30.1",
+    "rimraf": "6.0.1",
     "typescript": "5.8.3"
   },
   "packageManager": "pnpm@10.13.1",
